+++ clients/admin-ui/src/features/datamap/datamap.slice.ts
@@ -10,6 +10,7 @@ import {
   SYSTEM_PRIVACY_DECLARATION_DATA_SUBJECTS_NAME,
   SYSTEM_PRIVACY_DECLARATION_DATA_USE_NAME,
 } from "~/features/datamap/constants";
+import { DATAMAP_GROUPING, Page_DatamapReport_ } from "~/types/api";
 
 export interface DataCategoryNode {
   value: string;
@@ -63,6 +64,45 @@ const DEPRECATED_COLUMNS = [
 // API endpoints
 const datamapApi = baseApi.injectEndpoints({
   endpoints: (build) => ({
+    getMinimalDatamapReport: build.query<
+      Page_DatamapReport_,
+      {
+        groupBy: DATAMAP_GROUPING;
+        pageIndex: number;
+        pageSize: number;
+        search: string;
+        dataUses?: string;
+        dataCategories?: string;
+        dataSubjects?: string;
+      }
+    >({
+      query: ({
+        groupBy,
+        pageIndex,
+        pageSize,
+        search,
+        dataUses,
+        dataCategories,
+        dataSubjects,
+      }) => {
+        let queryString = `page=${pageIndex}&size=${pageSize}&group_by=${groupBy}`;
+        if (dataUses) {
+          queryString += `&${dataUses}`;
+        }
+        if (dataCategories) {
+          queryString += `&${dataCategories}`;
+        }
+        if (dataSubjects) {
+          queryString += `&${dataSubjects}`;
+        }
+        if (search) {
+          queryString += `&search=${search}`;
+        }
+        return {
+          url: `plus/datamap/minimal?${queryString}`,
+        };
+      },
+    }),
     getDatamap: build.query<DatamapTableData, { organizationName: string }>({
       query: ({ organizationName }) => ({
         url: `plus/datamap/${organizationName}`,
@@ -111,7 +151,11 @@ const datamapApi = baseApi.injectEndpoints({
   }),
 });
 
-export const { useGetDatamapQuery, useLazyGetDatamapQuery } = datamapApi;
+export const {
+  useGetDatamapQuery,
+  useLazyGetDatamapQuery,
+  useGetMinimalDatamapReportQuery,
+} = datamapApi;
 
 export interface SettingsState {
   columns?: DatamapColumn[];